[NEP-15720] Revert lib,test structure
