+++ lib/haml-i18n-extractor/core-ext/hash.rb
