+++ lib/haml-i18n-extractor/version.rb
@@ -1,0 +1,7 @@
+module Haml
+  module I18n
+    class Extractor
+      VERSION = "0.6.0"
+    end
+  end
+end
