--- lib/i18n-extractor/extraction/tagging_writer.rb
@@ -1,18 +1,0 @@
-module I18n
-  class Extractor
-    class TaggingWriter
-
-      DB = ".tags.i18n-extractor"
-      def initialize
-        @file = File.open(DB, "a+")
-      end
-
-      def write(path, lineno)
-        tag = "#{path}:#{lineno}\n"
-        @file.write(tag)
-        @file.flush
-      end
-
-    end
-  end
-end
