--- lib/i18n-extractor/version.rb
@@ -1,5 +1,0 @@
-module I18n
-  class Extractor
-    VERSION = "0.6.0"
-  end
-end
