+++ test/support/nested/dir/actual_no_underscore_yml.yml
@@ -23,7 +23,7 @@ en:
     home:
       cdx:
         dev:
-          i18n-extractor:
+          haml-i18n-extractor:
             nested:
               dir:
                 nested-no-leading-underscore:
