+++ test/support/nested/dir/nested-no-leading-underscore.i18n-extractor.haml
@@ -5,18 +5,18 @@
     %li
       = link_to account_settings_path do
         %i.fa.fa-fw.fa-cog
-        %span.text= t('.home.cdx.dev.i18n-extractor.testnested.dir.nested-no-leading-underscore.account_settings')
+        %span.text= t('.home.cdx.dev.haml-i18n-extractor.testnested.dir.nested-no-leading-underscore.account_settings')
     %li
       = link_to tariffs_path do
         %i.fa.fa-fw.fa-calculator
-        %span.text= t('.home.cdx.dev.i18n-extractor.testnested.dir.nested-no-leading-underscore.tariffs')
+        %span.text= t('.home.cdx.dev.haml-i18n-extractor.testnested.dir.nested-no-leading-underscore.tariffs')
     %li
       = link_to tariff_path(@tariff) do
         %i.fa.fa-fw.fa-calculator
         %span.text= @tariff.friendly_title
     %li.active
       %i.fa.fa-fw.fa-plus
-      %span.text= t('.home.cdx.dev.i18n-extractor.testnested.dir.nested-no-leading-underscore.add_volume_charge')
+      %span.text= t('.home.cdx.dev.haml-i18n-extractor.testnested.dir.nested-no-leading-underscore.add_volume_charge')
 
   #pageTitle
     .pull-left
@@ -27,16 +27,16 @@
         .btn-group
           = link_to new_tariff_path, class: 'btn btn-default btn-sm' do
             %i.fa.fa-fw.fa-plus
-            %span.text= t('.home.cdx.dev.i18n-extractor.testnested.dir.nested-no-leading-underscore.new_tariff')
+            %span.text= t('.home.cdx.dev.haml-i18n-extractor.testnested.dir.nested-no-leading-underscore.new_tariff')
       %li
         .btn-group
           = link_to account_settings_path, class: 'btn btn-default btn-sm' do
             %i.fa.fa-fw.fa-cog
-            %span.text= t('.home.cdx.dev.i18n-extractor.testnested.dir.nested-no-leading-underscore.settings')
+            %span.text= t('.home.cdx.dev.haml-i18n-extractor.testnested.dir.nested-no-leading-underscore.settings')
 
   .row
     .col-md-3
-      = render partial: t('.home.cdx.dev.i18n-extractor.testnested.dir.nested-no-leading-underscore.account_settingsnavigation'), locals: { parent: t('.home.cdx.dev.i18n-extractor.testnested.dir.nested-no-leading-underscore.tariff') }
+      = render partial: t('.home.cdx.dev.haml-i18n-extractor.testnested.dir.nested-no-leading-underscore.account_settingsnavigation'), locals: { parent: t('.home.cdx.dev.haml-i18n-extractor.testnested.dir.nested-no-leading-underscore.tariff') }
 
     .col-md-9
-      = render t('.home.cdx.dev.i18n-extractor.testnested.dir.nested-no-leading-underscore.form')
+      = render t('.home.cdx.dev.haml-i18n-extractor.testnested.dir.nested-no-leading-underscore.form')
